Move removal of portal_css and portal_js to plone 5.2
